Removed redundant auth service of candidate portal.
